# src/api/swagger-ui/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Parking Finder Backend API playground",
    "version": "1.0.0",
    "title": "Parking Finder Server"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["https", "http"],
  "paths": {
...
    "/user/verifyToken": {
      "post": {
        "tags": ["Forgot Password"],
        "summary": "Verify Password Reset Token",
        "description": "Verify that the provided password reset token is valid.",
        "operationId": "verifyToken",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "JSON object containing the password reset token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VerifyTokenRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset token is valid"
          },
          "400": {
            "description": "Invalid request parameters or invalid password reset token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "VerifyTokenRequest": {
      "type": "object",
      "properties": {
        "resetToken": {
          "type": "string"
        }
      },
      "required": ["resetToken"]
    },
...
  },
  "securityDefinitions": {
    "jwt": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your JWT token in the format 'Bearer {token}'."
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}